@@ -10,7 +10,7 @@ description: Demonstrates how to use the install_plugin plugin.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,7 +26,7 @@ dev_dependencies:
   install_plugin:
     path: ../
 
-  permission_handler: ^3.0.0
+  permission_handler: ^7.1.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
